+++ src/ploomber/entry.py
@@ -1,5 +1,37 @@
 """
-Warning: this code is highly experimental
+[Experimental module]
+
+The entry module creates command line interfaces for functions that return a
+DAG object
+
+Example:
+
+Assume module.sub_module.entry_point is a function that returns a DAG.
+
+Run this will parse the function parameters:
+
+    $ python -m ploomber.entry module.sub_module.entry_point --help
+
+To build the dag returned by the entry point:
+
+    $ python -m ploomber.entry module.sub_module.entry_point build
+
+
+To start an interactive session by loading the entry point first:
+
+    $ python -i -m ploomber.entry module.sub_module.entry_point status
+
+ipython -i -m also works. Once the interactive session starts, the object
+returned by the entry point will be available in the "dag" variable.
+
+Features:
+
+    * Parse docstring (numpydoc) and show it using --help
+    * Enable logging to standard output using --log LEVEL
+    * Pass function parameters using --PARAM_NAME VALUE
+    * If the function is decorated with @with_env, replace env variables using
+        --env__KEY VALUE
+
 """
 
 # TODO: print enviornment content on help and maybe on any other command
@@ -115,7 +147,11 @@ def main():
         # TODO: add a way of test this by the parameters it will use to
         # call the function, have an aux function to get those then another
         # to execute, test using the first one
-        print(getattr(entry(**{**kwargs, **replaced}), args.action)())
+        obj = entry(**{**kwargs, **replaced})
+
+        print(getattr(obj, args.action)())
+
+        return obj
 
 
 def flatten_dict(d, prefix=''):
@@ -134,4 +170,4 @@ def flatten_dict(d, prefix=''):
 
 
 if __name__ == '__main__':
-    main()
+    dag = main()