+++ src/ploomber/sources/sources.py
@@ -39,6 +39,8 @@ from ploomber.sql import infer
 
 
 class Source(abc.ABC):
+    """Source abstract class
+    """
 
     def __init__(self, value):
         self.value = Placeholder(value)
@@ -51,6 +53,7 @@ class Source(abc.ABC):
     def render(self, params):
         self.value.render(params)
         self._post_render_validation(self.value.value, params)
+        return self
 
     @property
     def loc(self):
@@ -58,6 +61,9 @@ class Source(abc.ABC):
 
     def __str__(self):
         return str(self.value)
+
+    def __repr__(self):
+        return "{}({})".format(type(self).__name__, self.value._value_repr)
 
     # required by subclasses
     @property