+++ tests/test_dag_task_reports.py
@@ -4,8 +4,8 @@ from pathlib import Path
 import pytest
 
 from ploomber.dag import DAG
-from ploomber.tasks import ShellScript, PythonCallable, SQLDump
-from ploomber.products import File
+from ploomber.tasks import ShellScript, PythonCallable, SQLScript
+from ploomber.products import File, SQLiteRelation
 from ploomber.constants import TaskStatus, DAGStatus
 from ploomber.exceptions import DAGBuildError, DAGRenderError
 
@@ -53,3 +53,19 @@ def test_dag_report_after_building(tmp_directory, executor):
     assert len(report['Elapsed (s)']) == 2
     assert len(report['name']) == 2
     assert len(report['Percentage']) == 2
+
+
+def test_dag_reports_sub_select_cols(sqlite_client_and_tmp_dir):
+    client, _ = sqlite_client_and_tmp_dir
+    dag = DAG()
+
+    dag.clients[SQLScript] = client
+    dag.clients[SQLiteRelation] = client
+
+    PythonCallable(touch_root, File('some_file'), dag, name='task')
+    sql = 'CREATE TABLE {{product}} AS SELECT * FROM data'
+    SQLScript(sql, SQLiteRelation(('data2', 'table')),
+              dag, name='task2')
+
+    assert dag.status()[['name', 'type']]
+    assert dag.build()[['Ran?', 'Elapsed (s)']]