+++ tests/test_sources.py
@@ -2,13 +2,90 @@ import pytest
 from unittest.mock import Mock
 
 from ploomber.exceptions import SourceInitializationError
-from ploomber.sources import SQLQuerySource, SQLScriptSource
+from ploomber.sources import SQLQuerySource, SQLScriptSource, GenericSource
 from ploomber.tasks import SQLScript
 from ploomber.products import SQLiteRelation
 from ploomber import DAG
-from ploomber.tasks import SQLScript
-from ploomber.products import PostgresRelation
+from ploomber.products import GenericSQLRelation
 from ploomber.sql import infer
+
+
+def test_generic_source_unrendered():
+    s = GenericSource('some {{placeholder}}')
+    assert repr(s) == 'GenericSource(some {{placeholder}})'
+
+    s = GenericSource(('some {{placeholder}} and some other large text '
+                       'with important details that we should not include'))
+    assert repr(s) == ('GenericSource(some {{placeholder}} and some other '
+                       'large text with important details that we...)')
+
+    s = GenericSource('some {{placeholder}}\nand some other large text ')
+    assert repr(s) == 'GenericSource(some {{placeholder}}...)'
+
+
+def test_generic_source_rendered():
+    s = GenericSource('some {{placeholder}}')
+    s.render({'placeholder': 'placeholder'})
+    assert repr(s) == 'GenericSource(some placeholder)'
+
+    s = GenericSource(('some {{placeholder}} and some other large text '
+                       'with important details that we should not include'))
+    s.render({'placeholder': 'placeholder'})
+    assert repr(s) == ('GenericSource(some placeholder and some other '
+                       'large text with important details that we sho...)')
+
+    s = GenericSource('some {{placeholder}}\nand some other large text ')
+    s.render({'placeholder': 'placeholder'})
+    assert repr(s) == 'GenericSource(some placeholder...)'
+
+
+@pytest.mark.parametrize('class_',
+                         [SQLScriptSource, SQLQuerySource])
+def test_sql_repr_unrendered(class_):
+    name = class_.__name__
+
+    # short case
+    s = class_('SELECT * FROM {{product}}')
+    assert repr(s) == name+'(SELECT * FROM {{product}})'
+
+    # long case, adds ...
+    s = class_(("SELECT * FROM {{product}} WHERE "
+                "some_very_very_long_column "
+                "= 'some_very_large_value'"))
+    assert repr(s) == name+("(SELECT * FROM {{product}} "
+                            "WHERE some_very_very_long_column "
+                            "= 'some_very_large...)")
+
+    # multi line case, adds ...
+    s = class_(("SELECT * FROM {{product}}\nWHERE "
+                "some_very_very_long_column"
+                " = 'some_very_large_value'"))
+    assert repr(s) == name+'(SELECT * FROM {{product}}...)'
+
+
+@pytest.mark.parametrize('class_',
+                         [SQLScriptSource, SQLQuerySource])
+def test_sql_repr_rendered(class_):
+    name = class_.__name__
+    render_arg = {'product': GenericSQLRelation(('schema', 'name', 'table'))}
+    s = class_('SELECT * FROM {{product}}')
+    s.render(render_arg)
+
+    assert repr(s) == name+'(SELECT * FROM schema.name)'
+
+    s = class_(("SELECT * FROM {{product}} WHERE "
+                "some_very_very_long_column "
+                "= 'some_very_large_value'"))
+    s.render(render_arg)
+    assert repr(s) == name+("(SELECT * FROM schema.name "
+                            "WHERE some_very_very_long_column "
+                            "= 'some_very_large...)")
+
+    s = class_(("SELECT * FROM {{product}}\nWHERE "
+                "some_very_very_long_column"
+                " = 'some_very_large_value'"))
+    s.render(render_arg)
+    assert repr(s) == name+'(SELECT * FROM schema.name...)'
 
 
 def test_can_parse_sql_docstring():